+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Add `EI` (EdgeInsets) shortcuts
+- Add `OB` (OutlinedButton) shortcuts
+
 ## 0.0.2
 
 Split some codes to [halo_dart](https://github.com/halowang/halo_dart)